clean up, refactor encapsulations
- merge PinotFunctionRegistry with FunctionRegistry
- renamed to match calcite.schema and calcite.sql from
  pinot.common.function package
